[form-factors] Move ``BSZ2015`` parametrization to separate files
